+++ nylas/client/client.py
@@ -72,6 +72,13 @@ def _validate(response):
     return response
 
 
+def _validate_availability_query(query):
+    if (query.get("emails", None) is None or len(query["emails"]) == 0) and (
+        query.get("calendars", None) is None or len(query["calendars"]) == 0
+    ):
+        raise ValueError("Must set either 'emails' or 'calendars' in the query.")
+
+
 class APIClient(json.JSONEncoder):
     """API client for the Nylas API."""
 
@@ -270,7 +277,7 @@ class APIClient(json.JSONEncoder):
         _validate(resp).json()
         return resp.json()
 
-    def free_busy(self, emails, start_at, end_at):
+    def free_busy(self, emails, start_at, end_at, calendars=None):
         if isinstance(emails, six.string_types):
             emails = [emails]
         if isinstance(start_at, datetime):
@@ -287,6 +294,10 @@ class APIClient(json.JSONEncoder):
             "start_time": start_time,
             "end_time": end_time,
         }
+        if calendars is not None and len(calendars) > 0:
+            data["calendars"] = calendars
+
+        _validate_availability_query(data)
         resp = self._request(HttpMethod.POST, url, json=data, cls=Calendar)
         _validate(resp)
         return resp.json()
@@ -321,6 +332,7 @@ class APIClient(json.JSONEncoder):
         round_robin=None,
         free_busy=None,
         open_hours=None,
+        calendars=None,
     ):
         if isinstance(emails, six.string_types):
             emails = [emails]
@@ -359,7 +371,10 @@ class APIClient(json.JSONEncoder):
             data["round_robin"] = round_robin
         if event_collection_id is not None:
             data["event_collection_id"] = event_collection_id
+        if calendars is not None and len(calendars) > 0:
+            data["calendars"] = calendars
 
+        _validate_availability_query(data)
         resp = self._request(HttpMethod.POST, url, json=data, cls=Calendar)
         _validate(resp)
         return resp.json()
@@ -374,10 +389,11 @@ class APIClient(json.JSONEncoder):
         buffer=None,
         free_busy=None,
         open_hours=None,
+        calendars=None,
     ):
         if isinstance(emails, six.string_types):
             emails = [[emails]]
-        elif isinstance(emails[0], list) is False:
+        elif len(emails) > 0 and isinstance(emails[0], list) is False:
             raise ValueError("'emails' must be a list of lists.")
         if isinstance(duration, timedelta):
             duration_minutes = int(duration.total_seconds() // 60)
@@ -412,7 +428,10 @@ class APIClient(json.JSONEncoder):
         }
         if buffer is not None:
             data["buffer"] = buffer
+        if calendars is not None and len(calendars) > 0:
+            data["calendars"] = calendars
 
+        _validate_availability_query(data)
         resp = self._request(HttpMethod.POST, url, json=data, cls=Calendar)
         _validate(resp)
         return resp.json()